change amount to int

+++ MainProject/Data/FoodFormModel.cs
@@ -3,12 +3,12 @@
     public class FoodFormModel
     {
         public int? VegetableId { get; set; }
-        public double? VegetableAmount { get; set; }
+        public int? VegetableAmount { get; set; }
         public int? FruitsId { get; set; }
-        public double? FruitsAmount { get; set; }
+        public int? FruitsAmount { get; set; }
         public int? MeatId { get; set; }
-        public double? MeatAmount { get; set; }
+        public int? MeatAmount { get; set; }
         public int? OtherId { get; set; }
-        public double? OtherAmount { get; set; }
+        public int? OtherAmount { get; set; }
     }
 }
